Towards a storlann/seosibh theme of DSpace.
Set up theme directory ("storlann") and created themed navbar/header.

Note that the cookies component (shared/cookies) did not implement i18n
so src/app was modified. (TODO see if this can be done as part of the theme.)

Note that the search-navbar component was not themeable so src/app was
modified. (TODO see if this can be done as part of the theme.)